requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asttokens
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
coloredlogs==15.0.1
comm
dataclasses-json==0.6.5
datasets==2.19.2
decorator
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
exceptiongroup
executing
fastapi==0.110.3
filelock==3.14.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipykernel
ipython
jedi
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client
jupyter_core
kubernetes==29.0.0
langchain==0.2.5
langchain-chroma==0.1.1
langchain-community==0.2.5
langchain-core==0.2.9
langchain-google-genai==1.0.6
langchain-text-splitters==0.2.1
langsmith==0.1.81
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
neo4j==5.20.0
nest_asyncio
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.25.1
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.2
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso
pexpect
pickleshare
pillow==10.3.0
platformdirs
posthog==3.5.0
prompt-toolkit
proto-plus==1.23.0
protobuf==4.25.3
psutil
ptyprocess
pure-eval
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments
py2neo
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq
regex==2024.4.28
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rouge==1.0.1
rouge_score==0.1.2
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
shellingham==1.5.4
six
sniffio==1.3.1
SQLAlchemy==2.0.29
stack-data
starlette==0.37.2
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
tornado
tqdm==4.66.2
traitlets
transformers==4.40.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp